Added in several other SiteMapDownloader classes and another recipe parser that looks for json+ld tags common to lots of sites and loads recipe data form a json object. Also made the setup processes for the start_id completely different. It may be good to refactor that out although not delete because the quality of the current recipe collection is unknown

+++ recipe_scraper/tools/sitemap_downloader.py
@@ -8,6 +8,7 @@ from aiohttp import ClientSession, TCPConnector
 from abc import ABCMeta
 from collections import OrderedDict
 from xml.parsers.expat import ExpatError
+from recipe_scraper.recipe_parsers import HRecipeParser, JsonLdParser
 
 SITEMAP_PATTERN = re.compile(r'(?<=sitemap:).+')
 
@@ -20,13 +21,20 @@ class SiteMapDownloader:
     recipe_url_pattern = None
     output_directory = None
     subdirectory_output = None
-    site_set = None
+    parser = None
+    site_set = set()
 
     completed_dir = 'collected'
 
     def __init__(self):
-        if any(not i for i in
-               [self.robots_url, self.output_directory, self.recipe_url_pattern, self.subdirectory_output]):
+        required_attributes = [
+            self.robots_url,
+            self.output_directory,
+            self.recipe_url_pattern,
+            self.subdirectory_output,
+            self.parser
+        ]
+        if any(not i for i in required_attributes):
             raise AssertionError("robots_url and recipe_url_pattern required for a sitemap url. Also set the output"
                                  "directory using the Abstract class set_download_directory")
         try:
@@ -38,11 +46,13 @@ class SiteMapDownloader:
             os.mkdir(os.path.join(self.output_directory, self.subdirectory_output, self.completed_dir))
         except OSError:
             pass
-        self.site_set = set()
         self._reverse = False
 
     async def get_sitemaps(self):
-        sitemaps = await self._find_sitemaps_from_robots(self.robots_url)
+        if isinstance(self.robots_url, str):
+            sitemaps = await self._find_sitemaps_from_robots(self.robots_url)
+        else:
+            sitemaps = self.robots_url
         while sitemaps:
             sitemap = sitemaps.pop()
             conn = TCPConnector(verify_ssl=False)
@@ -101,9 +111,12 @@ class SiteMapDownloader:
 
     def output_sitemap(self, xml, filename):
         output_file = os.path.join(self.output_directory, self.subdirectory_output, filename)
-        data = xmltodict.unparse(xml)
-        with gzip.open(output_file, 'wb') as f:
-            f.write(data.encode('utf-8'))
+        collected_output_file = os.path.join(self.output_directory, self.subdirectory_output,
+                                             self.completed_dir, filename)
+        if not os.path.isfile(output_file) and not os.path.isfile(collected_output_file):
+            data = xmltodict.unparse(xml)
+            with gzip.open(output_file, 'wb') as f:
+                f.write(data.encode('utf-8'))
 
     @property
     def get_links(self):
@@ -117,7 +130,7 @@ class SiteMapDownloader:
                         xml = xmltodict.parse(f.read())
                         links = self._search_sitemaps(xml, link_filter=self._recipe_link_filter)
                         for link in links:
-                            if self.site_set or link not in self.site_set:
+                            if link not in self.site_set:
                                 yield link
                     except ExpatError:
                         pass
@@ -167,6 +180,7 @@ class FoodSiteMapDownloader(SiteMapDownloader):
     robots_url = 'http://www.food.com/robots.txt'
     ignore_recipe_pattern = ['/review']
     recipe_url_pattern = ['www.food.com/recipe/']
+    parser = JsonLdParser.get_parser()
 
 
 class EpicuriousSiteMapDownloader(SiteMapDownloader):
@@ -175,6 +189,7 @@ class EpicuriousSiteMapDownloader(SiteMapDownloader):
     robots_url = 'http://www.epicurious.com/robots.txt'
     ignore_recipe_pattern = ['/review']
     recipe_url_pattern = ['www.epicurious.com/recipes/food/views']
+    parser = HRecipeParser.get_parser()
 
 
 class FoodnetworkSiteMapDownloader(SiteMapDownloader):
@@ -183,6 +198,7 @@ class FoodnetworkSiteMapDownloader(SiteMapDownloader):
     robots_url = 'http://www.foodnetwork.com/robots.txt'
     recipe_url_pattern = ['www.foodnetwork.com/recipes/']
     ignore_recipe_pattern = ['recipes/articles', 'recipes/photos', 'recipes/menus', 'recipes/packages']
+    parser = JsonLdParser.get_parser()
 
 
 class AllRecipesSiteMapDownloader(SiteMapDownloader):
@@ -190,6 +206,7 @@ class AllRecipesSiteMapDownloader(SiteMapDownloader):
     subdirectory_output = 'allrecipes'
     robots_url = 'http://allrecipes.com/robots.txt'
     recipe_url_pattern = ['allrecipes.com/recipe/']
+    parser = HRecipeParser.get_parser()
 
 
 class RecipeDepositorySiteMapDownloader(SiteMapDownloader):
@@ -197,3 +214,68 @@ class RecipeDepositorySiteMapDownloader(SiteMapDownloader):
     subdirectory_output = 'recipedepository'
     robots_url = 'http://www.therecipedepository.com/robots.txt'
     recipe_url_pattern = ['www.therecipedepository.com/recipe']
+    parser = HRecipeParser.get_parser()
+
+
+class SimplyRecipesSiteMapDownloader(SiteMapDownloader):
+    subdirectory_output = 'simply_recipes'
+    robots_url = 'http://www.simplyrecipes.com/robots.txt'
+    recipe_url_pattern = ['http://www.simplyrecipes.com/recipes/']
+    parser = HRecipeParser.get_parser()
+
+
+class BBCFoodSiteMapDownloader(SiteMapDownloader):
+    subdirectory_output = 'bbc_food'
+    robots_url = 'https://www.bbcgoodfood.com/robots.txt'
+    recipe_url_pattern = ['bbcgoodfood.com/recipes/']
+    parser = HRecipeParser.get_parser()
+
+
+class WillimasonomaSiteMapDownloader(SiteMapDownloader):
+    subdirectory_output = 'william_sonoma'
+    robots_url = 'http://www.williams-sonoma.com/robots.txt'
+    recipe_url_pattern = ['http://www.williams-sonoma.com/recipe']
+    parser = HRecipeParser.get_parser()
+
+
+class BonAppetiteSiteMapDownloader(SiteMapDownloader):
+    subdirectory_output = 'bon_appetite'
+    robots_url = 'http://www.bonappetit.com/robots.txt'
+    recipe_url_pattern = ['bonappetit.com/recipe/']
+    parser = JsonLdParser.get_parser()
+
+
+# No site map to generate links
+# class JaimeOliverSiteMapDownloader(SiteMapDownloader):
+#     subdirectory_output = 'jamie_oliver'
+#     robots_url = 'http://'
+#     recipe_url_pattern = ['jamieoliver.com/recipes/']
+#     parser = JsonLdParser.get_parser()
+
+
+class FineDiningSiteMapDownloader(SiteMapDownloader):
+    subdirectory_output = 'fine_dining'
+    robots_url = 'https://www.finedininglovers.com/robots.txt'
+    recipe_url_pattern = ['finedininglovers.com/recipes/']
+    parser = HRecipeParser.get_parser()
+
+
+class TheKitchnSiteMapDownloader(SiteMapDownloader):
+    subdirectory_output = 'thektichn'
+    robots_url = ['http://www.thekitchn.com/sitemap.xml']
+    recipe_url_pattern = ['thekitchn.com/recipe']
+    parser = HRecipeParser.get_parser()
+
+
+class ChowSiteMapDownloader(SiteMapDownloader):
+    subdirectory_output = 'chow'
+    robots_url = 'http://www.chowhound.com/robots.txt'
+    recipe_url_pattern = ['www.chowhound.com/recipes/']
+    parser = HRecipeParser.get_parser()
+
+
+class MyRecipeSiteMapDownloader(SiteMapDownloader):
+    subdirectory_output = 'myrecipes'
+    robots_url = ['http://www.myrecipes.com/sitemap-index.xml']
+    recipe_url_pattern = ['myrecipes.com/recipe/']
+    parser = JsonLdParser.get_parser()
